# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Investigative Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseStatus } from "./caseStatus";

export interface Case {
  id: number;
  title: string;
  description?: string;
  status: CaseStatus;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}