# lib/api-zod/src/generated/types/caseSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Investigative Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from "./case";
import type { Document } from "./document";
import type { Entity } from "./entity";
import type { Event } from "./event";
import type { MoneyFlow } from "./moneyFlow";
import type { Note } from "./note";
import type { Relationship } from "./relationship";
import type { TimelineEntry } from "./timelineEntry";

export interface CaseSummary {
  case: Case;
  entities: Entity[];
  documents: Document[];
  timeline: TimelineEntry[];
  events: Event[];
  notes: Note[];
  relationships: Relationship[];
  moneyFlows: MoneyFlow[];
}