# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Investigative Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  title: string;
  filePath?: string;
  source?: string;
  publishDate?: string;
  uploadedAt: string;
  caseId?: number;
}