# lib/api-zod/src/generated/types/moneyFlow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Investigative Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MoneyFlow {
  id: number;
  sourceEntityId: number;
  destinationEntityId: number;
  sourceEntityName?: string;
  destinationEntityName?: string;
  amount: number;
  date?: string;
  description?: string;
  supportingDocumentId?: number;
  caseId?: number;
  createdAt: string;
}