# lib/api-zod/src/generated/types/updateCaseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Investigative Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCaseInputStatus } from "./updateCaseInputStatus";

export interface UpdateCaseInput {
  title?: string;
  description?: string;
  status?: UpdateCaseInputStatus;
  tags?: string[];
}